# docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Appointment API - نظام حجز المواعيد",
        "description": "توثيق API لحجز المواعيد بين المرضى والممارسين والخدمات.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/appointments": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get all appointments",
                "operationId": "09fd012ab97514481bde57bbeac2699b",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Appointments"
                ],
                "summary": "إنشاء موعد جديد",
                "operationId": "15017f04a457b26e2a870c8a011a407a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointment"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تم إنشاء الموعد بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Chosen date and time is already booked"
                    }
                }
            }
        },
        "/api/appointments/{id}": {
            "get": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Get appointment by ID",
                "operationId": "13b46cd03a1dae29a4c4dd6e756d37da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the appointment",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response - Appointment data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Update an existing appointment",
                "operationId": "a779a3a6175fee8a68b75461718a1ec3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Appointment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Appointment updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Appointment"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Appointments"
                ],
                "summary": "Delete an appointment",
                "operationId": "cc5727c249d22f6719923c09625213e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Appointment ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Appointment deleted successfully"
                    },
                    "404": {
                        "description": "Appointment not found"
                    }
                }
            }
        },
        "/api/swagger-test": {
            "get": {
                "tags": [
                    "Swagger Test"
                ],
                "summary": "Simple Swagger test endpoint",
                "operationId": "0cdbe0219ee9a66d470c75e6bc8a6a9c",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Appointment": {
                "required": [
                    "patient_id",
                    "practitioner_id",
                    "service_id",
                    "appointment_date",
                    "appointment_time"
                ],
                "properties": {
                    "patient_id": {
                        "type": "integer"
                    },
                    "practitioner_id": {
                        "type": "integer"
                    },
                    "service_id": {
                        "type": "integer"
                    },
                    "appointment_date": {
                        "type": "string",
                        "format": "date"
                    },
                    "appointment_time": {
                        "type": "string",
                        "format": "time"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Appointments",
            "description": "API Endpoints for managing appointments"
        },
        {
            "name": "Swagger Test",
            "description": "Swagger Test"
        }
    ]
}